+++ src/main/scala/nl.knaw.dans.easy.deposit/DataFiles.scala
@@ -21,6 +21,7 @@ import java.nio.file.{ Path, Paths }
 import better.files._
 import nl.knaw.dans.bag.ChecksumAlgorithm.SHA1
 import nl.knaw.dans.bag.DansBag
+import nl.knaw.dans.easy.deposit.Errors.NoSuchFileInDepositException
 import nl.knaw.dans.easy.deposit.docs.FileInfo
 import nl.knaw.dans.lib.logging.DebugEnhancedLogging
