+++ src/main/scala/nl.knaw.dans.easy.deposit/docs/JsonUtil.scala
@@ -18,7 +18,7 @@ package nl.knaw.dans.easy.deposit.docs
 import java.lang.reflect.InvocationTargetException
 import java.nio.file.{ Path, Paths }
 
-import nl.knaw.dans.easy.deposit.BadRequestException
+import nl.knaw.dans.easy.deposit.Errors.InvalidDocumentException
 import nl.knaw.dans.easy.deposit.docs.StateInfo.State
 import nl.knaw.dans.easy.deposit.docs.dm._
 import org.json4s.Extraction.decompose
@@ -33,10 +33,6 @@ import scala.reflect.runtime.universe.typeOf
 import scala.util.{ Failure, Success, Try }
 
 object JsonUtil {
-
-  class InvalidDocumentException(document: String, t: Throwable = null)
-    extends BadRequestException(if (t == null) s"invalid $document"
-                                else s"invalid $document: ${ t.getMessage }")
 
   class PathSerializer extends CustomSerializer[Path](_ =>
     ( {
@@ -104,7 +100,7 @@ object JsonUtil {
         case cause: IllegalArgumentException if t.getMessage == "unknown error" => cause
         case _ => t
       }
-      Failure(new InvalidDocumentException(getType[A], cause))
+      Failure(InvalidDocumentException(getType[A], cause))
     }
 
     private def rejectNotExpectedContent[T](parsed: JValue, extracted: T): Try[Unit] = {
