+++ src/main/scala/nl.knaw.dans.easy.deposit/docs/StateInfo.scala
@@ -15,7 +15,7 @@
  */
 package nl.knaw.dans.easy.deposit.docs
 
-import nl.knaw.dans.easy.deposit.IllegalDepositStateException
+import nl.knaw.dans.easy.deposit.Errors.IllegalDepositStateException
 import nl.knaw.dans.easy.deposit.docs.JsonUtil.RichJsonInput
 import nl.knaw.dans.easy.deposit.docs.StateInfo.State.State
 import org.json4s.JsonInput
