+++ src/main/scala/nl.knaw.dans.easy.deposit/servlets/UserServlet.scala
@@ -15,9 +15,10 @@
  */
 package nl.knaw.dans.easy.deposit.servlets
 
+import nl.knaw.dans.easy.deposit.EasyDepositApiApp
+import nl.knaw.dans.easy.deposit.Errors.BadRequestException
 import nl.knaw.dans.easy.deposit.docs.JsonUtil._
 import nl.knaw.dans.easy.deposit.docs.UserInfo
-import nl.knaw.dans.easy.deposit.{ BadRequestException, EasyDepositApiApp }
 import nl.knaw.dans.lib.error._
 import nl.knaw.dans.lib.logging.servlet._
 import org.scalatra._
