+++ src/test/scala/nl.knaw.dans.easy.deposit/DataFilesSpec.scala
@@ -23,6 +23,7 @@ import java.util.UUID
 
 import better.files.StringOps
 import nl.knaw.dans.bag.v0.DansV0Bag
+import nl.knaw.dans.easy.deposit.Errors.NoSuchFileInDepositException
 import nl.knaw.dans.easy.deposit.docs.FileInfo
 import nl.knaw.dans.lib.error._
