+++ src/test/scala/nl.knaw.dans.easy.deposit/DepositDirSpec.scala
@@ -17,6 +17,7 @@ package nl.knaw.dans.easy.deposit
 
 import java.nio.file.attribute.PosixFilePermission
 
+import nl.knaw.dans.easy.deposit.Errors.{ CorruptDepositException, IllegalStateTransitionException }
 import nl.knaw.dans.easy.deposit.PidRequesterComponent.PidType.PidType
 import nl.knaw.dans.easy.deposit.docs.StateInfo
 import nl.knaw.dans.easy.deposit.docs.StateInfo.State
