+++ src/test/scala/nl.knaw.dans.easy.deposit/StagedFilesTargetSpec.scala
@@ -20,6 +20,7 @@ import java.nio.file.Paths
 
 import better.files.StringOps
 import nl.knaw.dans.bag.v0.DansV0Bag
+import nl.knaw.dans.easy.deposit.Errors.ConflictException
 import nl.knaw.dans.lib.error._
 
 import scala.util.{ Failure, Success }
