+++ src/test/scala/nl.knaw.dans.easy.deposit/docs/AgreementsSpec.scala
@@ -16,8 +16,8 @@
 package nl.knaw.dans.easy.deposit.docs
 
 import javax.xml.validation.Schema
+import nl.knaw.dans.easy.deposit.Errors.InvalidDocumentException
 import nl.knaw.dans.easy.deposit.TestSupportFixture
-import nl.knaw.dans.easy.deposit.docs.JsonUtil.InvalidDocumentException
 import nl.knaw.dans.easy.deposit.docs.dm.PrivacySensitiveDataPresent
 import org.joda.time.DateTime
 import org.xml.sax.SAXParseException
